cleanup components and fix information passing to order
