--- client/src/Components/Payment.js
@@ -1,67 +1,0 @@
-import React, { useState } from 'react'
-import { Form, Button, Col, Card, Container } from 'react-bootstrap'
-import { useDispatch, useSelector } from 'react-redux'
-import FormContainer from '../Components/FormContainer'
-import CheckoutSteps from '../Components/CheckoutSteps'
-import { savePaymentMethod } from '../Actions/cartActions'
-import { withRouter } from 'react-router-dom'
-
-const Payment = ({ history }) => {
-  const cart = useSelector((state) => state.cart)
-  const { shippingAddress } = cart
-
-  if (!shippingAddress.address) {
-    history.push('/shipping')
-  }
-
-  const [paymentMethod, setPaymentMethod] = useState('PayPal')
-
-  const dispatch = useDispatch()
-
-  const submitHandler = (e) => {
-    e.preventDefault()
-    dispatch(savePaymentMethod(paymentMethod))
-    history.push('/placeorder')
-  }
-
-  return (
-    <FormContainer>
-      <Card>
-        <Container>
-          <CheckoutSteps step1 step2 step3 />
-          <h1>Payment Method</h1>
-          <Form onSubmit={submitHandler}>
-            <Form.Group>
-              <Form.Label as='legend'>Select Method</Form.Label>
-              <Col>
-                <Form.Check
-                  type='radio'
-                  label='PayPal or Credit Card'
-                  id='PayPal'
-                  name='paymentMethod'
-                  value='PayPal'
-                  checked
-                  onChange={(e) => setPaymentMethod(e.target.value)}
-                ></Form.Check>
-                {/* <Form.Check
-              type='radio'
-              label='Stripe'
-              id='Stripe'
-              name='paymentMethod'
-              value='Stripe'
-              onChange={(e) => setPaymentMethod(e.target.value)}
-            ></Form.Check> */}
-              </Col>
-            </Form.Group>
-
-            <Button type='submit' variant='primary'>
-              Continue
-            </Button>
-          </Form>
-        </Container>
-      </Card>
-    </FormContainer>
-  )
-}
-
-export default withRouter(Payment)
