+++ server/controllers/orderController.js
@@ -45,7 +45,7 @@ const addOrderItems = asyncHandler(async (req, res) => {
 const getOrderById = asyncHandler(async (req, res) => {
   const order = await Order.findById(req.params.id).populate(
     'user',
-    'name email'
+    'name email number'
   )
 
   if (order) {
